Update AuditLogger.php
we've added a new EmailLoggingDriver class that implements a new log() method to send an email with the log message to the specified recipient. We've also updated the AuditLogger constructor to check for a new email driver type and create an instance of the EmailLoggingDriver class if that driver type is specified.